Added Read and Store
Added to two methods Read and Store to SimpleDB
